Create budget_calculator.py
This code uses Streamlit to create a web page with inputs for adding income and expenses, and displays the current balance and list of transactions. You can run this code using the Streamlit command-line tool to start a local web server, and the web page will be accessible in your web browser.